Hoist the shared script helpers out of processing/
_argparse.py, _dispatch.py and _scan.py are no longer processing-only --
the analysis layer composes them too -- so they move up to scripts/ and
every import becomes scripts._<helper>. The DEFAULT_* recipe/config
constants stay in scripts/processing/__init__.py.

Their three test modules combine into test_script_helpers.py, dropping
the `_dispatch as f` alias now that three helpers share the file, and
test_plotting_timeseries.py becomes test_timeseries_plotting.py.
