+++ scripts/processing/timeseries.py
@@ -43,6 +43,15 @@ from kpfpipe.utils.config import ConfigHandler
 from kpfpipe.utils.io import datecode_dirs_in_range
 from kpfpipe.utils.kpf import get_datecode, get_obs_id, is_datecode
 from kpfpipe.utils.logger import build_run_log_dir, setup_batch_logging
+from scripts._argparse import (
+    cache_parser,
+    data_dirs_parser,
+    logging_parser,
+    pool_parser,
+    resolve_dir_shortcuts,
+)
+from scripts._dispatch import _default_science_jobs, configure_runtime
+from scripts._scan import scan_datecodes, scan_night_to_cache
 from scripts.plotting.timeseries import PlotTimeseries
 from scripts.processing import (
     DEFAULT_MASTERS_CONFIG,
@@ -50,15 +59,6 @@ from scripts.processing import (
     DEFAULT_SCIENCE_CONFIG,
     DEFAULT_SCIENCE_RECIPE,
 )
-from scripts.processing._argparse import (
-    cache_parser,
-    data_dirs_parser,
-    logging_parser,
-    pool_parser,
-    resolve_dir_shortcuts,
-)
-from scripts.processing._dispatch import _default_science_jobs, configure_runtime
-from scripts.processing._scan import scan_datecodes, scan_night_to_cache
 
 logger = logging.getLogger(__name__)
