+++ tests/regression/test_timeseries_plotting.py
